archive/requirements.txt
bs4
pymongo
python-decouple